+++ web3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @direct.dev/web3
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [0df0ed9]
+  - @direct.dev/client@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes